add ability to process unknown residue types
Added the ability to process .pdb files in the case where the .pdb file contains some unknown residue type (i.e. the residue type specified in the .pdb is not in residuetypes.dat / not included in the default force field). In this case, the unknown residues will be cut out, and pdb2gmx will be called on a file containing just the known residues. After this, the internal list of unknown residues and the processed structure will be merge into a final .pdb file, and also the unknown residue types (and amounts) will be added to topol.top.